add using of primary constructor(and remove fields) to all classes

+++ RepairWorkshop.BLL/Services/RolePermissionService.cs
@@ -1,22 +1,19 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using RepairWorkshop.BLL.DTOs;
+using RepairWorkshop.BLL.Exceptions;
 using RepairWorkshop.BLL.Interfaces;
 using RepairWorkShop.DAL;
 using RepairWorkShop.DAL.Entities;
 using System;
 using System.Collections.Generic;
+using System.Security;
 using System.Text;
 
 namespace RepairWorkshop.BLL.Services
 {
-    public class RolePermissionService : IRolePermissionService
+    public class RolePermissionService(AppDbContext context) : IRolePermissionService
     {
-        public readonly AppDbContext _context;
-
-        public RolePermissionService(AppDbContext context)
-        {
-            _context = context;
-        }
+        public readonly AppDbContext _context = context;
 
         public async Task<RolePermission> CreateRolePermission(CreateRolePermissionDto dto)
         {
@@ -32,18 +29,27 @@ namespace RepairWorkshop.BLL.Services
             return rolePermission;
         }
 
-        public async Task DeleteRolePermission(int id)
+        public async Task DeleteRolePermission(int userRoleId, int permissionId)
         {
-            var rolePermission = await _context.RolePermissions.FindAsync(id);
+            var rolePermission = await _context.RolePermissions
+                .FirstOrDefaultAsync(r => r.UserRoleId == userRoleId &&
+                r.PermissionId == permissionId);
+
+            if (rolePermission == null)
+                throw new NotFoundException("role premission not found");
 
             _context.RolePermissions.Remove(rolePermission);
             await _context.SaveChangesAsync();
         }
 
-        // нема ід в нього, того не можна знайти одне.
-        public async Task<RolePermission> EditRolePermission(int id, EditRolePermissionDto dto)
+        public async Task<RolePermission> EditRolePermission(int userRoleId, int permissionId, EditRolePermissionDto dto)
         {
-            var rolePermission = await _context.RolePermissions.FindAsync(id);
+            var rolePermission = await _context.RolePermissions
+                .FirstOrDefaultAsync(r => r.UserRoleId == userRoleId &&
+                r.PermissionId == permissionId);
+
+            if (rolePermission == null)
+                throw new NotFoundException("role premission not found");
 
             rolePermission.UserRoleId = dto.UserRoleId;
             rolePermission.PermissionId = dto.PermissionId;
@@ -53,9 +59,11 @@ namespace RepairWorkshop.BLL.Services
             return rolePermission;
         }
 
-        public async Task<RolePermission> GetRolePermissionById(int id)
+        public async Task<RolePermission?> GetRolePermissionById(int userRoleId, int permissionId)
         {
-            return await _context.RolePermissions.FindAsync(id);
+            return await _context.RolePermissions
+                .FirstOrDefaultAsync(r => r.UserRoleId == userRoleId &&
+                r.PermissionId == permissionId);
         }
 
         public async Task<List<RolePermission>> GetAllRolePermissions()
